Added CS50 - Week 0 Problem Set 0 assignment
Objective: Discover the access code. 
How To Play:
When asked enter a number between 1-9 and press enter.
If correct, the screen will turn green, You will be asked to enter the next access code. 
You win after you guess all 3 access codes. 
You have a limit of 5 chances for each access code. 

I began with a few concepts. I wanted to do something that could be a bit easier to transfer to other languages. I figured asking for input, then responding to that input would be a fun way to show that off. 
The keys were added as a fun graphic to show the player's successful progress.